♻️ Refactor: 자동완성 input과 textfield input 분리
