+++ src/shared/ui/autocompleteInput/AutocompleteInput.type.ts
@@ -1,0 +1,5 @@
+import type { TextFieldInputProps } from '@shared/ui/text-field-input/TextFieldInput.type'
+
+export interface AutocompleteInputProps extends TextFieldInputProps {
+  suggestionList: string[]
+}
